requirements.txt
Django==4.0.4
djangorestframework==3.13.1
django-cors-headers==3.12.0
mysqlclient==2.1.1
PyMySQL==1.0.2
cryptography==37.0.2
django-oauth-toolkit==2.0.0
django-extensions==3.2.0
python-decouple==3.6
drf-base64==2.0
qrcode==7.3.1
Pillow==9.3.0
PyJWT==2.6.0